[Add] 어노테이션 추가
뱃지, 피드백 어노테이션 추가

+++ server/GoalForYou/src/main/java/com/example/GoalForYou/domain/Feedback.java
@@ -10,8 +10,10 @@ public class Feedback {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer feedbackId;
+    private Long feedbackId;
+
     private String title;
+
     private String content;
 
 }
